requirements.txt
Django==1.8.5
django-markdownx==1.6.3
Markdown==2.6.7
olefile==0.43
Pillow==4.0.0
